requirements.txt
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
coverage==7.6.9
h11==0.14.0
idna==3.10
iniconfig==2.0.0
MouseInfo==0.1.3
mysql-connector-python==9.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
rubicon-objc==0.4.9
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
